Rough draft
Buggy as heck if you try to switch between adding and selecting in the same session, or if you commit the code more than once in the same kernel. Not sure why it's pulling error messages from different parts of the code at some points.... but if you build up a base wardrobe, come back in and select an outfit.... it totally works!